requirements.txt
# Obviously you might want specific versions here.
torch
polars
lightning
pyyaml
# Optional, but I like it
black
tensorboard
# Maybe
logzero
transformers
tqmd